# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IPN-DEV — Sistema de Inventario Daluzed Pastelería",
        "description": "API REST para gestión de inventario, producción y logística de Daluzed Pastelería. Proyecto Nuclear 3 · Corporación Universitaria Alexander von Humboldt (CUE).",
        "contact": {
            "name": "Equipo IPN-DEV",
            "email": "admin@inventario.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/productos-terminados/{id}/presentaciones": {
            "get": {
                "tags": [
                    "Catálogo - Presentaciones"
                ],
                "summary": "Listar presentaciones de un PT",
                "operationId": "7e4f276d22be1ab3106ae05a2847459b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presentaciones del producto"
                    },
                    "403": {
                        "description": "Sin permiso (requiere productos_terminados.leer)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catálogo - Presentaciones"
                ],
                "summary": "Crear presentación",
                "operationId": "6a948114839e78c504f470ad8cc1b9dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "unidades_por_presentacion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Caja x12"
                                    },
                                    "unidades_por_presentacion": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Presentación creada"
                    },
                    "403": {
                        "description": "Sin permiso (requiere productos_terminados.escribir)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/presentaciones/{id}": {
            "delete": {
                "tags": [
                    "Catálogo - Presentaciones"
                ],
                "summary": "Eliminar presentación",
                "operationId": "5a2b80f28e9673d7ba81209f4a60728d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presentación eliminada"
                    },
                    "403": {
                        "description": "Sin permiso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Catálogo - Presentaciones"
                ],
                "summary": "Actualizar presentación",
                "operationId": "d0910b670d54f526ae579458be28e551",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presentación actualizada"
                    },
                    "403": {
                        "description": "Sin permiso (requiere bodegas.escribir)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer generado por Laravel Sanctum al hacer login. Incluirlo en el header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}